AutowiredFromBiz when bizVersion is black
